HexColorの分離
