+++ lib/View/terms.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import '../Widget/HexColor.dart';
 
 class TermsPage extends StatefulWidget {
   TermsPage({Key key, this.title}) : super(key: key);
@@ -277,17 +278,4 @@ class _TermsPageState extends State<TermsPage> {
       ),
     );
   }
-}
-
-//色をカラーコードで指定するためのクラス
-class HexColor extends Color {
-  static int _getColorFromHex(String hexColor) {
-    hexColor = hexColor.toUpperCase().replaceAll('#', '');
-    if (hexColor.length == 6) {
-      hexColor = 'FF' + hexColor;
-    }
-    return int.parse(hexColor, radix: 16);
-  }
-
-  HexColor(final String hexColor) : super(_getColorFromHex(hexColor));
 }
